requirements.txt

chromadb
langchain
tiktoken
Chroma
langchain-community
langchain_together
sentence-transformers
boto3
requests
transformers
datasets
faiss-cpu
python-docx
pandas
qdrant-client
pymilvus
weaviate-client
langchain-weaviate
selenium
streamlit
PyPDF2
pysqlite3-binary
undetected-chromedriver
webdriver-manager
pinecone[grpc]
undetected-chromedriver





